Add producer test (as part of receiver_test)
